+++ erdpy/delegation/staking_provider.py
@@ -1,11 +1,11 @@
 import binascii
 from pathlib import Path
 from typing import Any
-from os import path
 
-from erdpy.validators.validators_file import ValidatorsFile
 from erdpy import utils
+from erdpy.validators.validators_file import ValidatorsFile
 from erdpy.accounts import Account
+from erdpy.cli_password import load_password
 from erdpy.config import MetaChainSystemSCsCost
 from erdpy.validators.core import estimate_system_sc_call
 from erdpy.wallet.pem import parse_validator_pem
@@ -27,22 +27,25 @@ def prepare_args_for_create_new_staking_contract(args: Any):
 
 
 def prepare_args_for_add_nodes(args: Any):
-    validators_file = ValidatorsFile(Path(args.validators_file))
+    validators_file_path = Path(args.validators_file).expanduser()
+    validators_file = ValidatorsFile(validators_file_path)
 
     # TODO: Refactor, so that only address is received here.
     if args.using_delegation_manager:
         account = Account(address=args.delegation_contract)
     elif args.pem:
         account = Account(pem_file=args.pem)
-    elif args.keyfile and args.passfile:
-        account = Account(key_file=args.keyfile, pass_file=args.passfile)
+    elif args.keyfile:
+        password = load_password(args)
+        account = Account(key_file=args.keyfile, password=password)
 
     add_nodes_data = "addNodes"
     num_of_nodes = validators_file.get_num_of_nodes()
     for validator in validators_file.get_validators_list():
-        # get validator
-        validator_pem = validator.get("pemFile")
-        validator_pem = path.join(path.dirname(args.validators_file), validator_pem)
+        # Get path of "pemFile", make it absolute
+        validator_pem = Path(validator.get("pemFile")).expanduser()
+        validator_pem = validator_pem if validator_pem.is_absolute() else validators_file_path / validator_pem
+
         secret_key_bytes, bls_key = parse_validator_pem(validator_pem)
         signed_message = sign_message_with_bls_key(account.address.pubkey().hex(), secret_key_bytes.decode('ascii'))
         add_nodes_data += f"@{bls_key}@{signed_message}"