+++ erdpy/testnet/wallets.py
@@ -6,6 +6,7 @@ from typing import Dict, Tuple
 from erdpy import errors, utils
 from erdpy.accounts import Account
 from erdpy.wallet import pem
+from erdpy.workstation import get_tools_folder
 
 MAX_NUM_NODES = 12
 
@@ -74,11 +75,11 @@ def _get_observers_folder():
 
 
 def get_users() -> Dict[str, Account]:
-    result = OrderedDict()
+    result: Dict[str, Account] = OrderedDict()
 
     for pem_file in sorted(utils.list_files(_get_users_folder(), ".pem")):
         nickname = Path(pem_file).stem
-        account = Account(pem_file=pem_file)
+        account = Account(pem_file=str(pem_file))
         result[nickname] = account
 
     return result
@@ -89,4 +90,4 @@ def _get_users_folder():
 
 
 def _get_folder():
-    return Path(__file__).parent.absolute().joinpath("wallets")
+    return get_tools_folder() / "testwallets" / "latest"