+++ erdpy/tests/shared.sh
@@ -9,8 +9,8 @@ echo "PYTHONPATH = ${PYTHONPATH}"
 
 ERDPY="python3.8 -m erdpy.cli"
 SANDBOX=./testdata-out/SANDBOX
-USERS=../testnet/wallets/users
-VALIDATORS=../testnet/wallets/validators
+USERS=~/elrondsdk/testwallets/latest/users
+VALIDATORS=~/elrondsdk/testwallets/latest/validators
 DENOMINATION="000000000000000000"
 PROXY="http://localhost:7950"
 CHAIN_ID="local-testnet"